FEBio2: Introduced FESurfaceConstraint as a base class for nonlinear constraints that require a surface definition (e.g. FEVolumeConstraint) and had to rename what was previously called FESurfaceConstraint to FEPeriodicSurfaceConstraint. Renamed the FESurfacePairInteraction to FESurfacePairConstraint and made the interface consistent with FENLConstraint (I meant to derive this class from FENLConstraint but decided to hold off on pushing that through).
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@10001 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FECore/DataRecord.cpp
@@ -9,7 +9,7 @@
 UnknownDataField::UnknownDataField(const char* sz)
 {
 	m_szdata[0] = 0;
-	int l = strlen(sz);
+	int l = (int)strlen(sz);
 	if (l > 63) l = 63;
 	if (l>0) { strncpy(m_szdata, sz, l); m_szdata[l] = 0; }
 }
